reducing code duplication
