+++ packages/sdk-rtl/src/baseTransport.ts
@@ -24,7 +24,14 @@
 
  */
 
-import { StatusCode, addQueryParams } from './transport';
+import {
+  StatusCode,
+  addQueryParams,
+  agentPrefix,
+  mergeOptions,
+  LookerAppId,
+  sdkTimeout,
+} from './transport';
 import type {
   Authenticator,
   HttpMethod,
@@ -35,6 +42,7 @@ import type {
   RawObserver,
   SDKResponse,
   Values,
+  IRequestProps,
 } from './transport';
 
 export abstract class BaseTransport implements ITransport {
@@ -137,4 +145,59 @@ export abstract class BaseTransport implements ITransport {
   }
 
   abstract retry(wait: IRawRequest): Promise<IRawResponse>;
+
+  protected async initRequest(
+    method: HttpMethod,
+    path: string,
+    body?: any,
+    authenticator?: Authenticator,
+    options?: Partial<ITransportSettings>
+  ) {
+    const agentTag = options?.agentTag || this.options.agentTag || agentPrefix;
+    options = mergeOptions(
+      { ...this.options, ...{ headers: { [LookerAppId]: agentTag } } },
+      options ?? {}
+    );
+    const headers = options.headers ?? {};
+
+    // Make sure an empty body is undefined
+    if (!body) {
+      body = undefined;
+    } else {
+      if (typeof body !== 'string') {
+        body = JSON.stringify(body);
+        headers['Content-Type'] = 'application/json';
+      }
+    }
+
+    const ms = sdkTimeout(options) * 1000;
+    let signaller = AbortSignal.timeout(ms);
+    if ('signal' in options && options.signal) {
+      // AbortSignal.any may not be available, tolerate its absence
+      if (AbortSignal.any) {
+        signaller = AbortSignal.any([signaller, options.signal]);
+      } else {
+        console.warn('AbortSignal.any is not available in this transport');
+      }
+    }
+
+    let props: IRequestProps = {
+      body,
+      credentials: 'same-origin',
+      headers,
+      method,
+      url: path,
+      signal: signaller,
+    };
+
+    if (authenticator) {
+      // Add authentication information to the request
+      props = await authenticator(props);
+    }
+
+    // Transform to HTTP request headers at the last second
+    // props.headers = new Headers(props.headers) as any;
+
+    return props;
+  }
 }
